+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/AccountTransaction.java
@@ -32,6 +32,15 @@ public final class AccountTransaction {
         return buffer.array();
     }
 
+    /**
+     * Returns the type of the {@link Payload} associated
+     * with this `AccountTransaction`.
+     * @return the {@link com.concordium.sdk.transactions.Payload.PayloadType}
+     */
+    public Payload.PayloadType getType() {
+        return this.payload.getType();
+    }
+
     BlockItem toBlockItem() {
         return BlockItem.from(this);
     }